Finish most of APIC initialization + more Korlib classes added (#918)

+++ Source/Mosa.Korlib/System/Marvin.cs
@@ -117,7 +117,7 @@ namespace System
                 // running in a 64-bit process.
  
                 p0 += Unsafe.ReadUnaligned<uint>(ref data);
-                uint nextUInt32 = Unsafe.ReadUnaligned<uint>(ref Unsafe.AddByteOffset(ref data, 4));
+                uint nextUInt32 = Unsafe.ReadUnaligned<uint>(ref Unsafe.AddByteOffset(ref data, (IntPtr)4));
  
                 // One block round for each of the 32-bit integers we just read, 2x rounds total.
  
@@ -132,7 +132,7 @@ namespace System
                 // Requires https://github.com/dotnet/runtime/issues/6794 to be addressed first
                 // before we can realize the full benefits of this.
  
-                data = ref Unsafe.AddByteOffset(ref data, 8);
+                data = ref Unsafe.AddByteOffset(ref data, (IntPtr)8);
             } while (--loopCount > 0);
  
             // n.b. We've not been updating the original 'count' parameter, so its actual value is
@@ -168,7 +168,7 @@ namespace System
  
             // Read the last 4 bytes of the buffer.
  
-            uint partialResult = Unsafe.ReadUnaligned<uint>(ref Unsafe.Add(ref Unsafe.AddByteOffset(ref data, (nuint)count & 7), -4));
+            uint partialResult = Unsafe.ReadUnaligned<uint>(ref Unsafe.Add(ref Unsafe.AddByteOffset(ref data, (IntPtr)((uint)count & 7)), -4));
  
             // The 'partialResult' local above contains any data we have yet to read, plus some number
             // of bytes which we've already read from the buffer. An example of this is given below
@@ -234,7 +234,7 @@ namespace System
                 // [ AA          ]  -> 0x0000_80AA / 0xAA80_0000
                 // [ AA BB CC    ]  -> 0x0000_80CC / 0xCC80_0000
  
-                partialResult = Unsafe.AddByteOffset(ref data, (nuint)count & 2);
+                partialResult = Unsafe.AddByteOffset(ref data, (IntPtr)((uint)count & 2));
  
                 if (BitConverter.IsLittleEndian)
                 {
